cleanup fluent dir
