+++ fluent/BlockRune.ts
@@ -3,9 +3,9 @@ import { known } from "../rpc/mod.ts"
 import { ArrayRune, Rune } from "../rune/mod.ts"
 import { ValueRune } from "../rune/ValueRune.ts"
 import { hex } from "../util/mod.ts"
-import { Chain, ClientRune } from "./client.ts"
-import { CodecRune } from "./codec.ts"
-import { chain } from "./rpc_known_methods.ts"
+import { Chain, ClientRune } from "./ClientRune.ts"
+import { CodecRune } from "./CodecRune.ts"
+import { chain } from "./rpc_method_runes.ts"
 
 export class BlockRune<out U, out C extends Chain = Chain> extends Rune<known.SignedBlock, U> {
   constructor(_prime: BlockRune<U, C>["_prime"], readonly client: ClientRune<U, C>) {
