+++ fluent/ConstRune.ts
@@ -1,6 +1,6 @@
 import * as M from "../frame_metadata/mod.ts"
 import { Rune, ValueRune } from "../rune/mod.ts"
-import { PalletRune } from "./pallet.ts"
+import { PalletRune } from "./PalletRune.ts"
 
 export class ConstRune<out U> extends Rune<M.Constant, U> {
   constructor(_prime: ConstRune<U>["_prime"], readonly pallet: PalletRune<U>) {
