+++ fluent/InBlockBlockRune.ts
@@ -1,5 +1,5 @@
 import { BlockRune } from "./BlockRune.ts"
-import { Chain, ClientRune } from "./client.ts"
+import { Chain, ClientRune } from "./ClientRune.ts"
 
 export class InBlockBlockRune<out U, out C extends Chain = Chain> extends BlockRune<U, C> {
   constructor(_prime: InBlockBlockRune<U, C>["_prime"], client: ClientRune<U, C>) {
