+++ fluent/MetadataRune.ts
@@ -1,9 +1,9 @@
 import * as M from "../frame_metadata/mod.ts"
 import { Rune, RunicArgs, ValueRune } from "../rune/mod.ts"
 import { DeriveCodec, Ty } from "../scale_info/mod.ts"
-import { Chain, ClientRune } from "./client.ts"
-import { CodecRune } from "./codec.ts"
-import { PalletRune } from "./pallet.ts"
+import { Chain, ClientRune } from "./ClientRune.ts"
+import { CodecRune } from "./CodecRune.ts"
+import { PalletRune } from "./PalletRune.ts"
 
 export class MetadataRune<out U, out C extends Chain = Chain> extends Rune<M.Metadata, U> {
   constructor(_prime: MetadataRune<U>["_prime"], readonly client: ClientRune<U, C>) {
