+++ fluent/rpc_method_runes.ts
@@ -1,6 +1,6 @@
 import { known } from "../rpc/mod.ts"
 import * as U from "../util/mod.ts"
-import { rpcCall, rpcSubscription } from "./rpc.ts"
+import { rpcCall, rpcSubscription } from "./rpc_runes.ts"
 
 // TODO: generate the following?
 export namespace state {
