+++ providers/frame/mod.ts
@@ -1,7 +1,7 @@
 // moderate
 
-export * from "./Base.ts"
-export * from "./PolkadotDev.ts"
-export * from "./ProxyBase.ts"
-export * from "./Wss.ts"
-export * from "./Zombienet.ts"
+export * from "./Base.ts";
+export * from "./PolkadotDev.ts";
+export * from "./ProxyBase.ts";
+export * from "./Wss.ts";
+export * from "./Zombienet.ts";
