+++ rune/mod.ts
@@ -1,9 +1,9 @@
 // moderate --exclude *.test.ts
 
-export * from "./ArrayRune.ts"
-export * from "./Id.ts"
-export * from "./MetaRune.ts"
-export * from "./Rune.ts"
-export * from "./RunicArgs.ts"
-export * from "./Timeline.ts"
-export * from "./ValueRune.ts"
+export * from "./ArrayRune.ts";
+export * from "./Id.ts";
+export * from "./MetaRune.ts";
+export * from "./Rune.ts";
+export * from "./RunicArgs.ts";
+export * from "./Timeline.ts";
+export * from "./ValueRune.ts";
